Moved robots parse and site map functionality into robots.py out of scraper.py
