+++ .gitignore
@@ -111,3 +111,5 @@ venv.bak/
 
 # git
 .git/
+
+Logs/
